adding refactor to separate node only usage utils
