+++ packages/mdctl-api-driver/package.json
@@ -28,6 +28,7 @@
     "@medable/mdctl-api": "^1.0.27",
     "@medable/mdctl-cli": "^1.0.27",
     "@medable/mdctl-core-utils": "^1.0.27",
+    "@medable/mdctl-node-utils": "^1.0.0",
     "@medable/mdctl-credentials-provider-keychain": "^1.0.27",
     "@medable/mdctl-credentials-provider-pouchdb": "^1.0.27",
     "async-iter-stream": "^1.2.0",
