+++ packages/mdctl-cli/lib/config.js
@@ -5,7 +5,8 @@ const cloneObject = require('clone'),
       sh = require('shelljs'),
       jsyaml = require('js-yaml'),
       { merge } = require('lodash'),
-      { loadJsonOrYaml, pathTo } = require('@medable/mdctl-core-utils'),
+      { pathTo } = require('@medable/mdctl-core-utils'),
+      { loadJsonOrYaml } = require('@medable/mdctl-node-utils'),
       { privatesAccessor } = require('@medable/mdctl-core-utils/privates')
 
 let Undefined
