+++ packages/mdctl-cli/tasks/sandbox.js
@@ -4,7 +4,7 @@ const _ = require('lodash'),
       jsyaml = require('js-yaml'),
       fs = require('fs'),
       { rString, isSet, stringToBoolean } = require('@medable/mdctl-core-utils/values'),
-      { loadJsonOrYaml } = require('@medable/mdctl-core-utils'),
+      { loadJsonOrYaml } = require('@medable/mdctl-node-utils'),
       Task = require('../lib/task'),
       sandbox = require('@medable/mdctl-sandbox')
