+++ packages/mdctl-core-utils/crypto.js
@@ -1,6 +1,4 @@
 const randomBytes = require('randombytes'),
-      createHash = require('create-hash'),
-      fs = require('fs'),
       { rInt, rString } = require('./values'),
       alphaNumChars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789',
       allChars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789 !"#$%&\'()*+,-./:;<=>?@[\\]^_`{|}~'
@@ -55,30 +53,10 @@ function randomAlphaNumSym(length) {
 }
 
 
-function md5FileHash(filename) {
-  const size = 8192,
-        fd = fs.openSync(filename, 'r'),
-        hash = createHash('md5'),
-        buffer = Buffer.alloc(size)
-
-  try {
-    let bytesRead
-
-    do {
-      bytesRead = fs.readSync(fd, buffer, 0, size)
-      hash.update(buffer.slice(0, bytesRead))
-    } while (bytesRead === size)
-  } finally {
-    fs.closeSync(fd)
-  }
-
-  return hash.digest('hex')
-}
 
 module.exports = {
   secureRandomInt,
   randomAlphaNum,
   randomAlphaNumSym,
-  randomChars,
-  md5FileHash
+  randomChars
 }
