+++ packages/mdctl-core/index.js
@@ -1,9 +1,6 @@
-const { Manifest, ARegex } = require('./manifest'),
-      Fault = require('./fault')
+const Fault = require('./fault')
 
 
 module.exports = {
-  Manifest,
-  ARegex,
   Fault
 }
