+++ packages/mdctl-export-adapter-tree/index.js
@@ -3,12 +3,11 @@ const { Writable } = require('stream'),
       jp = require('jsonpath'),
       _ = require('lodash'),
       globby = require('globby'),
-      slugify = require('slugify'),
       request = require('request'),
       { Fault } = require('@medable/mdctl-core'),
       { stringifyContent } = require('@medable/mdctl-core-utils/values'),
-      { md5FileHash } = require('@medable/mdctl-core-utils/crypto'),
-      { ensureDir } = require('@medable/mdctl-core-utils/directory'),
+      { md5FileHash } = require('@medable/mdctl-node-utils/crypto'),
+      { ensureDir } = require('@medable/mdctl-node-utils/directory'),
       { privatesAccessor } = require('@medable/mdctl-core-utils/privates'),
       KNOWN_FILES = {
         assets: 'env/**/*.{ico,png,jpeg,jpg,gif,html,txt,bin,js}',
