+++ packages/mdctl-export-adapter-tree/package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@medable/mdctl-core": "^1.0.27",
     "@medable/mdctl-core-utils": "^1.0.27",
+    "@medable/mdctl-node-utils": "^1.0.0",
     "globby": "^10.0.1",
     "jsonpath": "^1.0.0",
     "lodash": "^4.17.14",
