+++ packages/mdctl-import-adapter/index.js
@@ -7,7 +7,7 @@ const EventEmitter = require('events'),
       pluralize = require('pluralize'),
       { ImportSection } = require('@medable/mdctl-core/streams/section'),
       { parseString, isCustomName } = require('@medable/mdctl-core-utils/values'),
-      { md5FileHash } = require('@medable/mdctl-core-utils/crypto'),
+      { md5FileHash } = require('@medable/mdctl-node-utils/crypto'),
       { privatesAccessor } = require('@medable/mdctl-core-utils/privates'),
       { OutputStream } = require('@medable/mdctl-core/streams/chunk-stream'),
       { Fault } = require('@medable/mdctl-core'),
