+++ packages/mdctl-import-adapter/package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@medable/mdctl-core": "^1.0.27",
     "@medable/mdctl-core-utils": "^1.0.27",
+    "@medable/mdctl-node-utils": "^1.0.0",
     "clone": "^2.1.2",
     "globby": "^9.1.0",
     "jsonpath": "^1.0.2",
