+++ packages/mdctl-manifest/test/manifest.js
