+++ packages/mdctl-node-utils/.npmignore
@@ -1,0 +1,23 @@
+*.log
+npm-debug.log*
+
+# Coverage directory used by tools like istanbul
+coverage
+.nyc_output
+
+# Dependency directories
+node_modules
+
+# npm package lock
+package-lock.json
+yarn.lock
+
+# project files
+test
+examples
+lint
+CHANGELOG.md
+.travis.yml
+.editorconfig
+.eslintrc
+.gitignore
