+++ packages/mdctl-node-utils/index.js
@@ -1,0 +1,17 @@
+const fs = require('fs'),
+      jsyaml = require('js-yaml'),
+      path = require('path')
+
+async function loadJsonOrYaml(file, multi) {
+  if (path.extname(file) === '.yaml') {
+    const docs = []
+    jsyaml.safeLoadAll(fs.readFileSync(file, 'utf8'), d => docs.push(d), { filename: file })
+    return multi ? docs : docs[0] || {}
+  }
+  return JSON.parse(fs.readFileSync(file, 'utf8'))
+}
+
+
+module.exports = {
+  loadJsonOrYaml
+}
