+++ packages/mdctl-node-utils/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@medable/mdctl-manifest",
-  "version": "1.0.27",
-  "description": "Medable Developer Tools :: Manifest Manager",
+  "name": "@medable/mdctl-node-utils",
+  "version": "1.0.0",
+  "description": "Medable Developer Node Tools :: Utils",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Medable/mdctl.git"
@@ -25,10 +25,8 @@
     "lint": "eslint *.js test"
   },
   "dependencies": {
-    "@medable/mdctl-core": "^1.0.27",
-    "@medable/mdctl-core-schemas": "^1.0.27",
-    "@medable/mdctl-core-utils": "^1.0.27",
-    "@medable/mdctl-node-utils": "^1.0.0"
+    "js-yaml": "^3.13.1",
+    "create-hash": "^1.2.0"
   },
   "devDependencies": {
     "chai": "^4.1.2",
@@ -42,6 +40,5 @@
     "istanbul": "^1.0.0-alpha",
     "mocha": "^5.0.0",
     "sinon": "^7.2.2"
-  },
-  "gitHead": "4d50c0a1a7e84d29ac2490436643434205ce3f6d"
+  }
 }
